participant personal

+++ libcms/apps/participants/administration/views.py
@@ -466,8 +466,14 @@ def library_user_list(request, managed_libraries=[]):
     })
 
 
+@login_required
 @transaction.atomic()
+@permission_required_or_403('participants.add_userlibrary')
+@decorators.must_be_org_user
 def add_library_user(request, library_id, managed_libraries=[]):
+    managed_libray_ids = [managed_library.id for managed_library in managed_libraries]
+    if managed_libray_ids and library_id not in managed_libray_ids:
+        raise HttpResponseForbidden(u'Вы не можете обслуживать эту организацию')
     library = get_object_or_404(models.Library, id=library_id)
     roles_queryset =  Group.objects.filter(name__startswith='role_')
     if request.method == 'POST':
@@ -509,10 +515,16 @@ def add_library_user(request, library_id, managed_libraries=[]):
         'user_roles_from': user_roles_from
     })
 
+@login_required
 @transaction.atomic()
+@permission_required_or_403('participants.change_userlibrary')
+@decorators.must_be_org_user
 def edit_library_user(request, id, managed_libraries=[]):
     # library = get_object_or_404(models.Library, id=library_id)
-    user_library = models.UserLibrary.objects.get(id=id)
+    user_library = get_object_or_404(models.UserLibrary, id=id)
+    managed_libray_ids = [managed_library.id for managed_library in managed_libraries]
+    if managed_libray_ids and user_library.library_id not in managed_libray_ids:
+        raise HttpResponseForbidden(u'Вы не можете обслуживать эту организацию')
     library = user_library.library
     if request.method == 'POST':
         user_form = forms.UserForm(request.POST, prefix='user_form', instance=user_library.user)
@@ -551,9 +563,15 @@ def edit_library_user(request, id, managed_libraries=[]):
         'user_roles_from': user_roles_from
     })
 
+@login_required
 @transaction.atomic()
-def delete_library_user(request, id):
-    user_library = models.UserLibrary.objects.get(id=id)
+@permission_required_or_403('participants.delete_userlibrary')
+@decorators.must_be_org_user
+def delete_library_user(request, id, managed_libraries=[]):
+    user_library = get_object_or_404(models.UserLibrary, id=id)
+    managed_libray_ids = [managed_library.id for managed_library in managed_libraries]
+    if managed_libray_ids and user_library.library_id not in managed_libray_ids:
+        raise HttpResponseForbidden(u'Вы не можете обслуживать эту организацию')
     user_library.user.delete()
     user_library.delete()
     return redirect('participants:administration:detail', id=user_library.library_id)
